Refactor large files for better maintainability
Split overly large files into focused, single-responsibility modules:

**tools/api.py (627→86 lines, -86% reduction):**
- openapi_spec.py: OpenAPI specification generation utilities
- http_client.py: HTTP client wrappers and request handling
- schema_resolver.py: Schema resolution and Pydantic model creation
- tool_factory.py: Tool construction and factory methods
- Maintains full backward compatibility through re-exports

**routes/v0/server.py (453→159 lines, -65% reduction):**
- server_schemas.py: Pydantic models and response schemas
- server_validation.py: Server configuration validation logic
- server_service.py: Business logic for server CRUD operations
- Clean separation of concerns with service layer pattern

**routes/v0/tool.py (407→207 lines, -49% reduction):**
- tool_schemas.py: Tool-related request/response models
- tool_services.py: Service classes for Arcade, MCP, A2A integrations
- Improved modularity for different tool service integrations

Benefits:
- Enhanced maintainability through single responsibility principle
- Improved code organization and readability
- Better testability with focused modules
- Preserved API compatibility and existing functionality
- Cleaner separation between routes, services, and schemas
